Re-arrange the connect methods.
Unfortunately, rdoc is pretty stupid and doesn't know how
to reach across files to generate documentation for methods
that are in different files from where the class is defined.
Because of that, we weren't generating the documentation for
all of the lookup-style and create-style methods (which are
mostly based off the connection object).  To appease rdoc,
move any methods hanging off of the connection object into
connect.c

Signed-off-by: Chris Lalancette <clalance@redhat.com>