Various dom.scheduler_parameter fixes.
First, use a macro to ease compatibility with the memory
parameters that will be implemented later.  Second, in
scheduler_type, use a function with rb_protect to make the
code easier to read.  Finally, in scheduler_parameter and
scheduler_parameter=, make sure to not leak params if an
exception occurs.

Signed-off-by: Chris Lalancette <clalance@redhat.com>