Revamp {scheduler,memory,blkio}_{get,set}_parameters.
This revamp was done with the realization that all of
these functions do very similar things, and hence we were
duplicating a lot of code.  Additionally, we were using
a couple of pretty big macros that would be nice to get
rid of.

Therefore, this revamp replaces all of that with a couple
of common functions (internal_{get,set}) that implement
the 90% of shared functionality.  The parts that are different
are handled via 3 callbacks, which each of the implementations
does differently.  This saves us over 100 lines of code,
and makes the functionality of all of them more consistent.

Signed-off-by: Chris Lalancette <clalance@redhat.com>